moved metadata package up one dir
