+++ pkg/metadata/common/payload.go
