+++ pkg/metadata/gohai/gohai_test.go
