+++ pkg/metadata/gohai/payload.go
