+++ pkg/metadata/host/host_darwin.go
