+++ pkg/metadata/host/host_darwin_test.go
