+++ pkg/metadata/host/host_linux_test.go
