+++ pkg/metadata/host/host_test.go
