+++ pkg/metadata/host/host_windows.go
