+++ pkg/metadata/host/host_windows_test.go
