+++ pkg/metadata/resources/resources_test.go
