+++ pkg/metadata/v5/v5_test.go
